chore(refactor): #1160 moved dialog folder under shared module
