+++ web/src/app/shared/shared.module.ts
@@ -9,10 +9,20 @@ import { AppMaterialModule } from '../app-material.module';
 import { ReactiveFormsModule } from '@angular/forms';
 import { PipesModule } from '../pipes/pipes.module';
 import { MarkdownDirective } from './directives/markdown.directive';
+import { DialogConfirmationComponent } from './dialog/confirmation/dialog-confirmation.component';
+import { DialogListComponent } from './dialog/list/dialog-list.component';
+import { DialogMarkdownComponent } from './dialog/markdown/dialog-markdown.component';
 
 @NgModule({
-    declarations: [MarkdownDirective],
+    declarations: [MarkdownDirective, DialogConfirmationComponent,
+        DialogListComponent,
+        DialogMarkdownComponent],
     imports: [CommonModule, FlexLayoutModule, AppMaterialModule, ReactiveFormsModule, PipesModule],
-    exports: [FlexLayoutModule, AppMaterialModule, ReactiveFormsModule, PipesModule, MarkdownDirective]
+    exports: [FlexLayoutModule, AppMaterialModule, ReactiveFormsModule, PipesModule, MarkdownDirective],
+    entryComponents: [
+        DialogConfirmationComponent,
+        DialogListComponent,
+        DialogMarkdownComponent
+    ]
 })
 export class SharedModule { }
